requirements.txt
requests
beautifulsoup4
lxml
opensearch-py
python-dotenv
sentence-transformers